Replace NXRange with mask in NewRegMatchField
For NXReg, OvS will accept value/mask on arbitary bitwise. If using
range[start:end], it is hard to combine discontinuous value/mask,
such as, to merge reg0[0:0] with reg0[2:2], with range method,
we will pass NXRange[0:2] and it will generate mask 0x7, but actually,
the mask is 0x5.

Here, we can use mask directly. But need the caller to compute the
mask with method:
    mask := ((uint32(1) << (rng[1] - rng[0] + 1)) - 1) << rng[0]

Signed-off-by: Jinjun Gao <gjinjun@gmail.com>