Support arbitary bitwise register mask
For NXReg, OvS will accept value/mask on arbitary bitwise. If using
range[start:end], it is hard to combine discontinuous value/mask,
such as, to merge reg0[0:0] with reg0[2:2], with range method,
we will pass NXRange[0:2] and it will generate mask 0x7, but actually,
the mask is 0x5.

In this patch, we change merge function to support discontinuous mask
according to each reg piece range and then merge all pieces for one register.

Signed-off-by: Gao Jinjun <gjinjun@gmail.com>