Replace contiv/libovsdb with ovn-org/libovsdb
contiv/libovsdb is abandoned and we were using a version from 2017

Instead we use github.com/ovn-org/libovsdb, which is still under active
development and is used by ovn-kubernetes.

We also move the ovsdbDriver code to an internal package: it is only
used for integration tests and does not need to be part of the public
API. All the functions that were not used anywhere are also removed.

Bump up version to v0.13.0.

Fixes #62

Signed-off-by: Antonin Bas <abas@vmware.com>